pyproject.toml
[project]
name = "secret-validator-grunt"
version = "0.1.0"
description = "Agentic validation of GitHub Secret Scanning alerts."
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "python-dotenv>=1.0",
    "pydantic>=2,<3",
    "pydantic-settings>=2,<3",
    "pyyaml>=6",
    "rich>=14",
    "typer>=0.21",
    "github-copilot-sdk>=0.1.21",
    "ghapi>=1.0.10",
    "validate-secrets @ git+https://github.com/advanced-security/validate-secrets.git@0a813752d37c0117b067c48ca3267d4f47518998",
]

[project.optional-dependencies]
dev = [
    "coverage>=7.13.3",
    "pytest>=9",
    "pytest-asyncio>=1.3.0",
    "yapf>=0.40",
]

[project.scripts]
fmt = "secret_validator_grunt.cli_fmt:fmt_main"
fmt-check = "secret_validator_grunt.cli_fmt:fmt_check_main"
secret-validator-grunt = "secret_validator_grunt.main:entrypoint"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.metadata]
allow-direct-references = true

[tool.hatch.build.targets.wheel]
packages = ["src/secret_validator_grunt"]

[tool.hatch.build.targets.sdist]
include = ["src/secret_validator_grunt", "README.md", "LICENSE", ".env.example"]

[tool.pytest.ini_options]
testpaths = ["tests"]
addopts = "-q"
pythonpath = ["src"]
